Update xác minh tài khoan

+++ ShoeShop/Areas/Admin/Pages/user/Delete.cshtml.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.EntityFrameworkCore;
-using DataAccess.Models;
 using BusinessObject.user;
+using DataAccess.Models;
 
 namespace ShoeShop.Areas.Admin.Pages.user
 {
@@ -20,7 +16,7 @@ namespace ShoeShop.Areas.Admin.Pages.user
         }
 
         [BindProperty]
-      public User User { get; set; } = default!;
+        public User User { get; set; } = default!;
 
         public async Task<IActionResult> OnGetAsync(int? id)
         {
@@ -35,10 +31,17 @@ namespace ShoeShop.Areas.Admin.Pages.user
             {
                 return NotFound();
             }
-            else 
+            else
             {
                 User = user;
             }
+
+            if (!await CheckAccessAsync())
+            {
+                Redirect("AccesDenied");
+
+            }
+
             return Page();
         }
 
@@ -48,15 +51,34 @@ namespace ShoeShop.Areas.Admin.Pages.user
             {
                 return NotFound();
             }
+
             var user = await _userService.GetByIdAsync(id.Value);
 
             if (user != null)
             {
                 User = user;
+
+                if (!await CheckAccessAsync())
+                {
+                    return Forbid(); // Hoặc Redirect đến trang khác nếu không có quyền
+                }
+
                 await _userService.DeleteAsync(id.Value);
             }
 
             return RedirectToPage("./Index");
+        }
+
+        private async Task<bool> CheckAccessAsync()
+        {
+            var userSession = HttpContext.Session.GetString("UserSession");
+            if (string.IsNullOrEmpty(userSession))
+            {
+                return false;
+            }
+
+            var user = System.Text.Json.JsonSerializer.Deserialize<User>(userSession);
+            return user?.RoleId == 1; // Kiểm tra quyền truy cập
         }
     }
 }
